requirements.txt
absl-py==2.3.1
accelerate==1.10.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.1
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.4.0
beautifulsoup4==4.14.3
blinker==1.9.0
cachetools==6.2.3
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
colorlog==6.10.1
cryptography==46.0.3
datasets==4.5.0
dill==0.4.0
docling==2.69.1
docling-core==2.60.1
docling-ibm-models==3.10.3
docling-parse==4.7.3
et_xmlfile==2.0.0
evaluate==0.4.6
exceptiongroup==1.3.1
Faker==37.12.0
filelock==3.19.1
filetype==1.2.0
Flask==3.1.2
flask-cors==6.0.2
Flask-JWT-Extended==4.7.1
flatbuffers==25.12.19
frozenlist==1.8.0
fsspec==2025.10.0
gast==0.7.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.48.0
google-genai==1.47.0
google-pasta==0.2.0
grpcio==1.78.0
h11==0.16.0
h5py==3.14.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==0.36.2
idna==3.11
importlib_metadata==8.7.1
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
jsonlines==4.0.0
jsonref==1.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
keras==3.10.0
latex2mathml==3.78.1
libclang==18.1.1
lxml==6.0.2
Markdown==3.9
markdown-it-py==3.0.0
marko==2.2.2
MarkupSafe==3.0.3
mdurl==0.1.2
ml_dtypes==0.5.4
mpire==2.10.2
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.18
namex==0.1.0
narwhals==2.13.0
networkx==3.2.1
nltk==3.9.2
numpy==2.0.2
ocrmac==1.0.1 ; sys_platform == "darwin"
omegaconf==2.3.0
opencv-python==4.13.0.92
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.18.0
packaging==25.0
pandas==2.3.3
pdfminer.six==20251107
pdfplumber==0.11.8
pillow==11.3.0
plotly==6.5.0
pluggy==1.6.0
polyfactory==3.2.0
propcache==0.4.1
protobuf==6.33.2
psutil==7.2.2
pyarrow==21.0.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pyclipper==1.3.0.post6
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.11.0
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
PyJWT==2.10.1
pylatexenc==2.10
pyobjc-core==11.1; sys_platform == "darwin"
pyobjc-framework-Cocoa==11.1; sys_platform == "darwin"
pyobjc-framework-CoreML==11.1; sys_platform == "darwin"
pyobjc-framework-Quartz==11.1; sys_platform == "darwin"
pyobjc-framework-Vision==11.1; sys_platform == "darwin"
pypdf==6.6.2
pypdfium2==5.2.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-pptx==1.0.2
pytz==2025.2
PyYAML==6.0.3
rapidocr==3.6.0
referencing==0.36.2
regex==2026.1.15
requests==2.32.5
rich==14.3.2
rouge_score==0.1.2
rpds-py==0.27.1
rsa==4.9.1
rtree==1.4.1
safetensors==0.7.0
scipy==1.13.1
semchunk==2.2.2
sentencepiece==0.2.1
shapely==2.0.7
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
soupsieve==2.8.3
streamlit==1.50.0
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
termcolor==3.1.0
tokenizers==0.22.2
toml==0.10.2
torch==2.8.0
torchaudio==2.8.0
torchvision==0.23.0
tornado==6.5.3
tqdm==4.67.1
transformers==4.57.6
tree-sitter==0.23.2
tree-sitter-c==0.23.4
tree-sitter-javascript==0.23.1
tree-sitter-python==0.23.6
tree-sitter-typescript==0.23.2
typer==0.19.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.2
websockets==15.0.1
Werkzeug==3.1.5
wrapt==2.1.1
xlsxwriter==3.2.9
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
